requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
pandas==2.1.4
pandas-ta==0.3.14b
anthropic==0.18.1
python-dotenv==1.0.0
apscheduler==3.10.4
sqlalchemy==2.0.25
aiosqlite==0.19.0
colorlog==6.8.2
httpx==0.26.0
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
websockets==12.0
